+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 - Added support for [Solspace Calendar](https://solspace.com/craft/calendar/) specific contexts ([#36](https://github.com/angell-co/Spoon/issues/36))
+- Added partial support for [Craft Commerce](https://craftcms.com/commerce) product type contexts ([#49](https://github.com/angell-co/Spoon/issues/49))
 
 ### Changed
 - Moved the global block type context out of the plugin settings section ([#51](https://github.com/angell-co/Spoon/issues/51))